Very basic global configuration, needs fixes
- Global configuration using +all tag instead of a separate file
- Moved all the parsing logic into cli/parsing.py, leaving inside
  cli/__init__.py only the argparse code
- Updated README to reflect new changes
