+++ example/pipelines.yml
@@ -3,60 +3,66 @@
 # You can also look at pipelines_annotated.yml for an extensively annotated version of this file
 #
 
++all:
+  hooks:
+    on_job_start:
+      - common.UselessHook.whatever
+
 simple_task:
-    inputs:
-      - file.CsvInput:
-          directory: "./data"
+  inputs:
+    - file.CsvInput:
+        directory: "./data"
 
-    expose:
-      - file.CsvInput:
-        - "a_csv_file.csv": example_csv
-        - "a_csv_file.csv": example_csv
+  expose:
+    - file.CsvInput:
+      - "a_csv_file.csv": example_csv
+      - "a_csv_file.csv": example_csv
 
-    hooks:
-      on_job_start:
-        - common.UselessHook.whatever
-      on_pipeline_start:
-        - common.UselessHook.whatever
+  hooks:
+    on_job_start:
+      - common.UselessHook.whatever
+    on_pipeline_start:
+      - common.UselessHook.whatever
 
-    steps:
-      - common.PreProcessor
-      - MainTask: common.PreProcessor
-      - Later.print_something: MainTask
+  steps:
+    - common.PreProcessor
+    - MainTask: common.PreProcessor
+    - Later.print_something: MainTask
+
 
 sligthly_more_complex :
-    inputs:
-      - pgsql.PgSqlInput:
-          username: carlo
-          password:
-          host: localhost
-          port: 5432
-          database: postgres
-      - utils.DummyInput
-      - file.CsvInput:
-          directory: "./data"
+  inputs:
+    - pgsql.PgSqlInput:
+        username: carlo
+        password:
+        host: localhost
+        port: 5432
+        database: postgres
+    - utils.DummyInput
+    - file.CsvInput:
+        directory: "./data"
 
-    outputs:
-      - pgsql.PgSqlOutput:
-          username: carlo
-          password: !env PG_PWD
-          host: localhost
-          port: 5432
-          database: postgres
+  outputs:
+    - pgsql.PgSqlOutput:
+        username: carlo
+        password: !env PG_PWD
+        host: localhost
+        port: 5432
+        database: postgres
 
-    expose:
-      - PgSqlInput:
-        - dataset: dataset
-      - file.CsvInput:
-        - "a_csv_file.csv": example_csv
-      - utils.DummyInput:
-        - "whatever" : anempytdataframe
+  expose:
+    - PgSqlInput:
+      - dataset: dataset
+    - file.CsvInput:
+      - "a_csv_file.csv": example_csv
+    - utils.DummyInput:
+      - "whatever" : anempytdataframe
 
-    steps:
-      - common.PreProcessor
-      - PredictorA: common.PreProcessor
-      - PredictorB: common.PreProcessor
-      - PredictorC: common.PreProcessor
-      - Ensemble: [PredictorA, PredictorB, PredictorC]
-      - common.PostProcessor: Ensemble
+  steps:
+    - common.PreProcessor
+    - PredictorA: common.PreProcessor
+    - PredictorB: common.PreProcessor
+    - PredictorC: common.PreProcessor
+    - Ensemble: [PredictorA, PredictorB, PredictorC]
+    - common.PostProcessor: Ensemble
